apps/backend-ai/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymupdf
pymilvus
python-multipart
